# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monitoreo del hogar",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/casas": {
      "get": {
        "tags": [
          "casas"
        ],
        "summary": "Listar",
        "description": "Las casas del usuario, con su conexión y cuántas alarmas tienen abiertas.",
        "operationId": "listar_api_v1_casas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumenCasa"
                  },
                  "type": "array",
                  "title": "Response Listar Api V1 Casas Get"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "casas"
        ],
        "summary": "Crear",
        "description": "Registra una casa. Su central necesita además un usuario MQTT (scripts/usuario_mqtt.sh).",
        "operationId": "crear_api_v1_casas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CasaNueva"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Casa"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/casas/{casa_id}": {
      "get": {
        "tags": [
          "casas"
        ],
        "summary": "Ver",
        "operationId": "ver_api_v1_casas__casa_id__get",
        "parameters": [
          {
            "name": "casa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Casa Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Casa"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "casas"
        ],
        "summary": "Cambiar",
        "operationId": "cambiar_api_v1_casas__casa_id__patch",
        "parameters": [
          {
            "name": "casa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Casa Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioCasa"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Casa"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ajustes": {
        "properties": {
          "recordatorio_min": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Recordatorio Min"
          },
          "recordatorio_conexion_min": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Recordatorio Conexion Min"
          },
          "minutos_central_caida": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Minutos Central Caida"
          },
          "avisar_nodo_sin_conexion": {
            "type": "boolean",
            "title": "Avisar Nodo Sin Conexion"
          },
          "avisar_resueltas": {
            "type": "boolean",
            "title": "Avisar Resueltas"
          }
        },
        "type": "object",
        "required": [
          "recordatorio_min",
          "recordatorio_conexion_min",
          "minutos_central_caida",
          "avisar_nodo_sin_conexion",
          "avisar_resueltas"
        ],
        "title": "Ajustes"
      },
...
      "CambioCasa": {
        "properties": {
          "nombre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          }
        },
        "type": "object",
        "title": "CambioCasa"
      },
...
      "Casa": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "creada_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creada En"
          },
          "ajustes": {
            "$ref": "#/components/schemas/Ajustes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nombre",
          "creada_en",
          "ajustes"
        ],
        "title": "Casa"
      },
...
      "CasaNueva": {
        "properties": {
          "codigo": {
            "type": "string",
            "pattern": "^[a-z0-9-]{4,40}$",
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Nombre"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "nombre"
        ],
        "title": "CasaNueva"
      },
      "DetalleError": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "campos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campos"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "DetalleError"
      },
      "Error": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/DetalleError"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Error",
        "description": "{\"detail\": {\"codigo\": ..., \"mensaje\": ...}}"
      },
...
      "ResumenCasa": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "type": "string",
            "enum": [
              "admin",
              "cuidador"
            ],
            "title": "Rol"
          },
          "online": {
            "type": "boolean",
            "title": "Online"
          },
          "alarmas_abiertas": {
            "type": "integer",
            "title": "Alarmas Abiertas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "nombre",
          "rol",
          "online",
          "alarmas_abiertas"
        ],
        "title": "ResumenCasa"
      },
...
    }
  }
}